HW-8 data 모듈 생성 및 구현
