--- 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/api/ApiClient.kt
@@ -1,5 +1,0 @@
-package com.mtjin.androidarchitecturestudy.api
-
-object ApiClient {
-    const val BASE_URL = "https://openapi.naver.com/"
-}
