--- 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/data/login/source/LoginRepositoryImpl.kt
@@ -1,13 +1,0 @@
-package com.mtjin.androidarchitecturestudy.data.login.source
-
-import com.mtjin.androidarchitecturestudy.data.login.source.local.LoginLocalDataSource
-
-class LoginRepositoryImpl(private val loginLocalDataSource: LoginLocalDataSource) :
-    LoginRepository {
-
-    override var autoLogin: Boolean
-        get() = loginLocalDataSource.autoLogin
-        set(value) {
-            loginLocalDataSource.autoLogin = value
-        }
-}
