--- 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/data/search/Movie.kt
@@ -1,27 +1,0 @@
-package com.mtjin.androidarchitecturestudy.data.search
-
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.google.gson.annotations.SerializedName
-
-@Entity(tableName = "movie")
-data class Movie(
-    @SerializedName("actor")
-    val actor: String,
-    @SerializedName("director")
-    val director: String,
-    @SerializedName("image")
-    val image: String,
-    @SerializedName("link")
-    val link: String,
-    @SerializedName("pubDate")
-    val pubDate: String,
-    @SerializedName("subtitle")
-    val subtitle: String,
-    @PrimaryKey(autoGenerate = false)
-    @SerializedName("title")
-    val title: String,
-    @SerializedName("userRating")
-    val userRating: String
-)
