--- 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/data/search/MovieResponse.kt
@@ -1,17 +1,0 @@
-package com.mtjin.androidarchitecturestudy.data.search
-
-
-import com.google.gson.annotations.SerializedName
-
-data class MovieResponse(
-    @SerializedName("display")
-    val display: Int,
-    @SerializedName("items")
-    val movies: List<Movie>,
-    @SerializedName("lastBuildDate")
-    val lastBuildDate: String,
-    @SerializedName("start")
-    val start: Int,
-    @SerializedName("total")
-    val total: Int
-)
