--- 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/data/search/source/MovieRepository.kt
@@ -1,18 +1,0 @@
-package com.mtjin.androidarchitecturestudy.data.search.source
-
-import com.mtjin.androidarchitecturestudy.data.search.Movie
-
-interface MovieRepository {
-    fun getSearchMovies(
-        query: String,
-        success: (List<Movie>) -> Unit,
-        fail: (Throwable) -> Unit
-    )
-
-    fun getPagingMovies(
-        query: String,
-        start: Int,
-        success: (List<Movie>) -> Unit,
-        fail: (Throwable) -> Unit
-    )
-}
