--- 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/data/search/source/local/MovieDatabase.kt
@@ -1,12 +1,0 @@
-package com.mtjin.androidarchitecturestudy.data.search.source.local
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import com.mtjin.androidarchitecturestudy.data.search.Movie
-
-@Database(entities = [Movie::class], version = 1, exportSchema = false)
-abstract class MovieDatabase : RoomDatabase() {
-    abstract fun movieDao(): MovieDao
-}
