+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/java/com/mtjin/data/api/ApiClient.kt
@@ -1,0 +1,5 @@
+package com.mtjin.data.api
+
+object ApiClient {
+    const val BASE_URL = "https://openapi.naver.com/"
+}
