+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/java/com/mtjin/data/login/source/LoginRepository.kt
@@ -1,0 +1,6 @@
+package com.mtjin.data.login.source
+
+interface LoginRepository {
+
+    var autoLogin: Boolean
+}
