+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/java/com/mtjin/data/login/source/LoginRepositoryImpl.kt
@@ -1,0 +1,13 @@
+package com.mtjin.data.login.source
+
+import com.mtjin.data.login.source.local.LoginLocalDataSource
+
+class LoginRepositoryImpl(private val loginLocalDataSource: LoginLocalDataSource) :
+    LoginRepository {
+
+    override var autoLogin: Boolean
+        get() = loginLocalDataSource.autoLogin
+        set(value) {
+            loginLocalDataSource.autoLogin = value
+        }
+}
