+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/java/com/mtjin/data/login/source/local/LoginLocalDataSource.kt
@@ -1,0 +1,5 @@
+package com.mtjin.data.login.source.local
+
+interface LoginLocalDataSource {
+    var autoLogin: Boolean
+}
