+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/java/com/mtjin/data/login/source/local/LoginLocalDataSourceImpl.kt
@@ -1,0 +1,12 @@
+package com.mtjin.data.login.source.local
+
+import com.mtjin.data.utils.PreferenceManager
+
+class LoginLocalDataSourceImpl(private val preferenceManager: PreferenceManager) :
+    LoginLocalDataSource {
+    override var autoLogin: Boolean
+        get() = preferenceManager.autoLogin
+        set(value) {
+            preferenceManager.autoLogin = value
+        }
+}
