+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/java/com/mtjin/data/search/model/Movie.kt
@@ -1,0 +1,26 @@
+package com.mtjin.data.search.model
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import com.google.gson.annotations.SerializedName
+
+@Entity(tableName = "movie")
+data class Movie(
+    @SerializedName("actor")
+    val actor: String,
+    @SerializedName("director")
+    val director: String,
+    @SerializedName("image")
+    val image: String,
+    @SerializedName("link")
+    val link: String,
+    @SerializedName("pubDate")
+    val pubDate: String,
+    @SerializedName("subtitle")
+    val subtitle: String,
+    @PrimaryKey(autoGenerate = false)
+    @SerializedName("title")
+    val title: String,
+    @SerializedName("userRating")
+    val userRating: String
+)
