+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/java/com/mtjin/data/search/model/MovieResponse.kt
@@ -1,0 +1,15 @@
+package com.mtjin.data.search.model
+import com.google.gson.annotations.SerializedName
+
+data class MovieResponse(
+    @SerializedName("display")
+    val display: Int,
+    @SerializedName("items")
+    val movies: List<Movie>,
+    @SerializedName("lastBuildDate")
+    val lastBuildDate: String,
+    @SerializedName("start")
+    val start: Int,
+    @SerializedName("total")
+    val total: Int
+)
