+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/java/com/mtjin/data/search/source/MovieRepository.kt
@@ -1,0 +1,18 @@
+package com.mtjin.data.search.source
+
+import com.mtjin.data.search.model.Movie
+
+interface MovieRepository {
+    fun getSearchMovies(
+        query: String,
+        success: (List<Movie>) -> Unit,
+        fail: (Throwable) -> Unit
+    )
+
+    fun getPagingMovies(
+        query: String,
+        start: Int,
+        success: (List<Movie>) -> Unit,
+        fail: (Throwable) -> Unit
+    )
+}
