+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/java/com/mtjin/data/search/source/local/MovieDao.kt
@@ -1,0 +1,23 @@
+package com.mtjin.data.search.source.local
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import com.mtjin.data.search.model.Movie
+
+@Dao
+interface MovieDao {
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    fun insertMovies(movies: List<Movie>)
+
+    @Query("SELECT * FROM movie")
+    fun getAllMovies(): List<Movie>
+
+    @Query("SELECT * FROM movie WHERE title LIKE '%' || :title || '%'")
+    fun getMoviesByTitle(title: String): List<Movie>
+
+    @Query("DELETE FROM movie")
+    fun deleteAllMovies()
+}
