+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/java/com/mtjin/data/search/source/local/MovieLocalDataSource.kt
@@ -1,0 +1,10 @@
+package com.mtjin.data.search.source.local
+
+import com.mtjin.data.search.model.Movie
+
+interface MovieLocalDataSource {
+    fun insertMovies(movies: List<Movie>)
+    fun getAllMovies(): List<Movie>
+    fun getSearchMovies(title: String): List<Movie>
+    fun deleteAllMovies()
+}
