+++ 2003/mtjin/AndroidArchitectureStudy/data/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="app_name">data</string>
+</resources>
