+++ 2003/mtjin/AndroidArchitectureStudy/settings.gradle
@@ -1,2 +1,2 @@
-include ':app'
+include ':app', ':data'
 rootProject.name='AndroidArchitectureStudy'
